search pills fixup: Use grid to support pill overflow styling.
This is a fixup commit that should be merged into the main
search pill commit, but is separated for the ease of reviewing
the grid changes separately from other changes.

Before this commit, pill overflow was broken (sticking out
of the search box to the right) and now the search bar extends
vertically to allow pills to wrap when there are too many pills
for one line.